basic MX4cK side is done
I thought I committed this earlier, but I guess I didn't, so this is my
first commit.

For the sake of a basic communication demo, this program only spits
accelerometer data out to the PC when the PC sends data its way.  It is
a controlled response.

Next steps:
- Make the MX4cK program no longer able to receive data (or at least get
data out of the RX FIFO and dump it).
- Make the MX4cK program continually spit out accelerometer data and not
receive data.
- Make the PC program no longer able to transmit data.
- Make the PC program continually read accelerometer data and print it
to the console.