started adding openGL stuff
I had to fight with some stuff because I discovered that QT's GL window
widget includes some winsock stuff (at least I think it does; the
compiler was complaining about redefinitions).  The way that they
include it does not allow WinSock2.h to be included before
QtOpenGL/qglwidget.

So that took awhile to chase down.  I put a note in my code to remind me
of the necessity of this build order.

Note: I didn't implement any openGL stuff, but I included a bunch of
stuff, and I wanted to check that nothing was redefined in QT's GL
window widget.  The socket stuff still works.