finally got it to render; plus memory cleanup
One of the big things was that I had to set vertex attribute divisors
for the transformation matrix.

During the process, I added a shape generation capacity for a "double
triangle".  This was a simple drawing thing so that I could track from a
simple triangle to a cube.

Also added mouse-to-camera tracking as part of the debugging process.

Also added a bunch of real time updates in the camera's mouse tracking
handlers.

Next step: Rotate the cube my toying with it's "up" vector.