now have a cube that faces mouse
After a lot of frustration and experimentation and trying to grasp some
of the mathematics behind rotation and such (which I still don't get), I
managed to get a cube to track my mouse when I click and drag.

Yay!

Next step: Try to figure out how to introduce the socket communication.
The QT widget's exec() method is blocking.