started integrated QTcpSocket and signals
Aside from the last commit to clean up a little clutter in main, it has
been awhile.

I got to the point where I was ready for serial data, but then realized
that both QApplication::exec() and the serial reading were blocking.  To
get around that, I needed to do some multi-threading stuff.

Thankfully, Qt already handles that, but it took two days to get through
understanding how to integrate it with Visual Studio.  It isn't perfect,
and I still need to run to the meta compiler myself when I do something
new, but I'm alright for now.

Next step: Change the data being spit out of the microcontroller to
three floats (one per X, Y, and Z axis) instead of a string.