reworked rotation mechanism
Previously, I was rotating selectively around the horizontal axes (X and
Z), but I reworked it to calculate a rotation axis by taking the cross
product of the accelerometer data (effectively a unit vector that is
being treated as a directional vector relative to the object center) and
the world up vector.

That was a bit of a revelation, but it took a lot of work and tinkering
to figure out how the rotation vectors worked.  In the end, the solution
was, thankfully, rather simple.

The rotation calculation are in the my_shape class' "point thineself at
relative point" method.